chore(refactor): move python code to app directory
