+++ app/__init__.py
@@ -1,0 +1,1 @@
+"""Package de l'application demo-geocontext."""
